requirements.txt
Django>=6.0.1
djangorestframework>=3.16.0
python-dotenv>=1.0.0
requests>=2.32.0
